+++ prophet/prophet.go
@@ -1,10 +1,15 @@
 package prophet
 
 import (
+	"context"
 	"fmt"
 	"log/slog"
+	"sync"
+	"time"
 
 	lru "github.com/hashicorp/golang-lru/v2"
+
+	"github.com/cosmos/cosmos-sdk/client"
 )
 
 // queueBufferSize sets the default size for incoming queues, i.e. the number
@@ -21,6 +26,9 @@ type P struct {
 
 	resultsWriter *s[FutureResult]
 	votesWriter   *s[Vote]
+
+	selfAddressRwLock sync.RWMutex
+	selfAddress       []byte
 }
 
 // New returns an initialized P. Call [P.Run] to start the main loop.
@@ -47,7 +55,7 @@ func New() (*P, error) {
 //
 // Goroutines are started to execute incoming futures and verifying incoming
 // future results.
-func (p *P) Run() error {
+func (p *P) Run(tendermintRpc string) error {
 	futures, err := newDedupFutureReader(p.futures)
 	if err != nil {
 		return fmt.Errorf("failed to create futures dedup reader: %w", err)
@@ -65,6 +73,28 @@ func (p *P) Run() error {
 	if err := ExecVotes(proposals, p.votesWriter); err != nil {
 		return fmt.Errorf("failed to run votes loop: %w", err)
 	}
+
+	go func() {
+		client, err := client.NewClientFromNode(tendermintRpc)
+		if err != nil {
+			panic(err)
+		}
+		timeout := time.Now().Add(30 * time.Second)
+		for {
+			if time.Now().After(timeout) {
+				return
+			}
+			status, err := client.Status(context.Background())
+			if err != nil {
+				time.Sleep(100 * time.Millisecond)
+				continue
+			}
+			p.selfAddressRwLock.Lock()
+			defer p.selfAddressRwLock.Unlock()
+			p.selfAddress = status.ValidatorInfo.Address
+			return
+		}
+	}()
 
 	return nil
 }
@@ -94,6 +124,12 @@ func (p *P) Results() ([]FutureResult, func()) {
 	return values, func() {
 		p.resultsWriter.Remove(values...)
 	}
+}
+
+func (p *P) SelfAddress() []byte {
+	p.selfAddressRwLock.RLock()
+	defer p.selfAddressRwLock.RUnlock()
+	return p.selfAddress
 }
 
 // Votes returns a slice with all the votes of futures' results that have been