+++ warden/testutil/keeper/async.go
@@ -32,12 +32,16 @@ func AsyncKeeper(t testing.TB) (keeper.Keeper, sdk.Context) {
 	registry := codectypes.NewInterfaceRegistry()
 	cdc := codec.NewProtoCodec(registry)
 	authority := authtypes.NewModuleAddress(govtypes.ModuleName)
+	asyncModuleAddress := authtypes.NewModuleAddress(types.ModuleName)
 
 	k := keeper.NewKeeper(
 		cdc,
 		runtime.NewKVStoreService(storeKey),
 		log.NewNopLogger(),
 		authority.String(),
+		nil,
+		nil,
+		asyncModuleAddress,
 		nil,
 	)