+++ warden/x/async/keeper/keeper.go
@@ -2,14 +2,26 @@ package keeper
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"math/big"
 
 	"cosmossdk.io/collections"
 	"cosmossdk.io/core/store"
+	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/log"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+	evmosconf "github.com/evmos/evmos/v20/server/config"
+	evmkeeper "github.com/evmos/evmos/v20/x/evm/keeper"
+	evmostypes "github.com/evmos/evmos/v20/x/evm/types"
 
+	"github.com/warden-protocol/wardenprotocol/precompiles/callbacks"
+	precommon "github.com/warden-protocol/wardenprotocol/precompiles/common"
 	"github.com/warden-protocol/wardenprotocol/prophet"
 	types "github.com/warden-protocol/wardenprotocol/warden/x/async/types/v1beta1"
 )
@@ -22,11 +34,14 @@ type (
 
 		// the address capable of executing a MsgUpdateParams message. Typically, this
 		// should be the x/gov module account.
-		authority string
+		authority          string
+		asyncModuleAddress sdk.Address
 
-		futures  *FutureKeeper
-		handlers *HandlersKeeper
-		votes    collections.Map[collections.Pair[uint64, []byte], int32]
+		futures       *FutureKeeper
+		handlers      *HandlersKeeper
+		getEvmKeeper  func(_placeHolder int16) *evmkeeper.Keeper
+		accountKeeper types.AccountKeeper
+		votes         collections.Map[collections.Pair[uint64, []byte], int32]
 
 		p *prophet.P
 	}
@@ -50,6 +65,9 @@ func NewKeeper(
 	logger log.Logger,
 	authority string,
 	p *prophet.P,
+	getEvmKeeper func(_placeHolder int16) *evmkeeper.Keeper,
+	asyncModuleAddress sdk.Address,
+	accountKeeper types.AccountKeeper,
 	//selfValAddr sdk.ConsAddress,
 ) Keeper {
 	if _, err := sdk.AccAddressFromBech32(authority); err != nil {
@@ -74,14 +92,17 @@ func NewKeeper(
 	}
 
 	return Keeper{
-		cdc:          cdc,
-		storeService: storeService,
-		authority:    authority,
-		logger:       logger,
+		cdc:                cdc,
+		storeService:       storeService,
+		authority:          authority,
+		asyncModuleAddress: asyncModuleAddress,
+		logger:             logger,
 
-		futures:  futures,
-		handlers: handlers,
-		votes:    votes,
+		futures:       futures,
+		handlers:      handlers,
+		getEvmKeeper:  getEvmKeeper,
+		accountKeeper: accountKeeper,
+		votes:         votes,
 
 		p: p,
 	}
@@ -107,6 +128,10 @@ func (k Keeper) AddFutureResult(ctx context.Context, id uint64, submitter, outpu
 	}
 
 	if err := k.SetFutureVote(ctx, id, submitter, types.FutureVoteType_VOTE_TYPE_VERIFIED); err != nil {
+		return err
+	}
+
+	if err := k.futureReadyCallback(ctx, id, output); err != nil {
 		return err
 	}
 
@@ -145,6 +170,124 @@ func (k Keeper) GetFutureVotes(ctx context.Context, futureId uint64) ([]types.Fu
 	}
 
 	return votes, nil
+}
+
+func (k Keeper) futureReadyCallback(
+	ctx context.Context,
+	id uint64,
+	output []byte,
+) error {
+	future, err := k.futures.Get(ctx, id)
+
+	if err != nil {
+		return err
+	}
+
+	if future.Callback == "" {
+		return nil
+	}
+
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+
+	abi, err := callbacks.IAsyncCallbackMetaData.GetAbi()
+	if err != nil {
+		return err
+	}
+
+	method := "cb"
+	if _, ok := abi.Methods[method]; !ok {
+		return fmt.Errorf("invalid callback method: %v", method)
+	}
+
+	cbAddress, err := precommon.AddressFromBech32Str(future.Callback)
+
+	if err != nil {
+		return err
+	}
+
+	data, err := abi.Pack(method, id, output)
+	if err != nil {
+		return err
+	}
+
+	res, err := k.callEVMWithData(
+		sdkCtx,
+		common.BytesToAddress(k.asyncModuleAddress.Bytes()),
+		&cbAddress,
+		data,
+	)
+
+	if res.Failed() {
+		// Do not throw error if contract fails
+		return nil
+	}
+
+	return err
+}
+
+func (k Keeper) callEVMWithData(
+	ctx sdk.Context,
+	from common.Address,
+	contract *common.Address,
+	data []byte,
+) (*evmostypes.MsgEthereumTxResponse, error) {
+	fromAcc := k.accountKeeper.GetAccount(ctx, from.Bytes())
+	if fromAcc == nil {
+		fromAcc = k.accountKeeper.NewAccountWithAddress(ctx, from.Bytes())
+		k.accountKeeper.SetAccount(ctx, fromAcc)
+	}
+	nonce := fromAcc.GetSequence()
+
+	evmKeeper := k.getEvmKeeper(0)
+
+	args, err := json.Marshal(evmostypes.TransactionArgs{
+		From: &from,
+		To:   contract,
+		Data: (*hexutil.Bytes)(&data),
+	})
+	if err != nil {
+		return nil, errorsmod.Wrapf(errortypes.ErrJSONMarshal, "failed to marshal tx args: %s", err.Error())
+	}
+
+	gasRes, err := evmKeeper.EstimateGasInternal(ctx, &evmostypes.EthCallRequest{
+		Args:   args,
+		GasCap: evmosconf.DefaultGasCap,
+	}, evmostypes.Internal)
+	if err != nil {
+		return nil, err
+	}
+
+	// 1. Gas estimation also consumes gas.
+	// 2. Precompile creates new gas meter limited to contract gas cap. This new gas meter should consume gas from prev gas meter.
+	// 3. TODO: configurable gas limit on module level.
+	gasCap := gasRes.Gas + ctx.GasMeter().GasConsumed()
+
+	msg := ethtypes.NewMessage(
+		from,
+		contract,
+		nonce,
+		big.NewInt(0), // amount
+		gasCap,        // gasLimit
+		big.NewInt(0), // gasFeeCap
+		big.NewInt(0), // gasTipCap
+		big.NewInt(0), // gasPrice
+		data,
+		ethtypes.AccessList{}, // AccessList
+		false,                 // isFake
+	)
+
+	res, err := evmKeeper.ApplyMessage(ctx, msg, evmostypes.NewNoOpTracer(), true)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := fromAcc.SetSequence(fromAcc.GetSequence() + 1); err != nil {
+		return nil, err
+	}
+
+	k.accountKeeper.SetAccount(ctx, fromAcc)
+
+	return res, nil
 }
 
 type completedFuture struct {